+++ CHANGES.rst
@@ -15,6 +15,13 @@
 - Workaround the PyPy 7.1 JIT bug using MySQL Connector/Python. It is no
   longer necessary to disable the JIT in PyPy 7.1.
 
+- On PostgreSQL, use PostgreSQL's efficient binary ``COPY FROM`` to
+  store objects into the database. This can be 20-40% faster. See
+  :issue:`247`.
+
+- Silence a warning about ``cursor.connection`` from pg8000. See
+  :issue:`238`.
+
 3.0a2 (2019-06-19)
 ==================